+++ src/utils/index.ts
@@ -7,3 +7,4 @@ export * from './output';
 export * from './schema';
 export * from './string';
 export * from './warning';
+export * from './zod-to-ts-type';